Upload glazed CDN
glazed CDN is the Content from the oOP-CDN-Server (oop.ownonline.eu)

The aim is to show that the CDN is as secure and open (OIpenSource) as the player files provided here for viewing and forking. So you can be sure that the CDN executes the same code as provided here in the repository. The presentation page hosted on the CDN and standalone documentation is excluded from this. Please note that these are only the source files