First preparation for new test version 0.0.1.7 Gardenia
You're slowly getting your English!  This version will be the first with a basic language manager. For this, the browser and PHP will be configured for an automatic client-based or URL-parameter selected language via the parameter hl

The language management already works. The most time-consuming thing is to replace the German hard-coded with the language management queries.

To start with, only German and English will be available. However, the base is set and all UI elements must now be adapted accordingly. After that it's done. Then the player can also speak Chinese if you want it to.

Track progress in an executing environment on the CDN: https://oop.ownonline.eu/play/0-0-1-7/?lfmstream=radiomoordorf